+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/implementation/NonAzureOpenAIClientImpl.java
@@ -6,6 +6,7 @@ package com.azure.ai.openai.implementation;
 import com.azure.ai.openai.models.ChatCompletionsOptions;
 import com.azure.ai.openai.models.CompletionsOptions;
 import com.azure.ai.openai.models.EmbeddingsOptions;
+import com.azure.ai.openai.models.ModerationOptions;
 import com.azure.core.annotation.BodyParam;
 import com.azure.core.annotation.ExpectedResponses;
 import com.azure.core.annotation.HeaderParam;
@@ -195,6 +196,44 @@ public final class NonAzureOpenAIClientImpl {
             @HostParam("endpoint") String endpoint,
             @HeaderParam("accept") String accept,
             @BodyParam("application/json") BinaryData chatCompletionsOptions,
+            RequestOptions requestOptions,
+            Context context);
+
+        @Post("/moderations")
+        @ExpectedResponses({200})
+        @UnexpectedResponseExceptionType(
+            value = ClientAuthenticationException.class,
+            code = {401})
+        @UnexpectedResponseExceptionType(
+            value = ResourceNotFoundException.class,
+            code = {404})
+        @UnexpectedResponseExceptionType(
+            value = ResourceModifiedException.class,
+            code = {409})
+        @UnexpectedResponseExceptionType(HttpResponseException.class)
+        Mono<Response<BinaryData>> getModerations(
+            @HostParam("endpoint") String endpoint,
+            @HeaderParam("accept") String accept,
+            @BodyParam("application/json") BinaryData moderationsOptions,
+            RequestOptions requestOptions,
+            Context context);
+
+        @Post("/moderations")
+        @ExpectedResponses({200})
+        @UnexpectedResponseExceptionType(
+            value = ClientAuthenticationException.class,
+            code = {401})
+        @UnexpectedResponseExceptionType(
+            value = ResourceNotFoundException.class,
+            code = {404})
+        @UnexpectedResponseExceptionType(
+            value = ResourceModifiedException.class,
+            code = {409})
+        @UnexpectedResponseExceptionType(HttpResponseException.class)
+        Response<BinaryData> getModerationsSync(
+            @HostParam("endpoint") String endpoint,
+            @HeaderParam("accept") String accept,
+            @BodyParam("application/json") BinaryData moderationsOptions,
             RequestOptions requestOptions,
             Context context);
     }
@@ -715,6 +754,177 @@ public final class NonAzureOpenAIClientImpl {
             OPEN_AI_ENDPOINT,
             accept,
             chatCompletionsOptionsUpdated,
+            requestOptions,
+            Context.NONE);
+    }
+
+    /**
+     * Gets moderations for the provided input. The moderations endpoint is a tool you can use to check
+     * whether content complies with OpenAI's usage policies. Developers can thus identify content
+     * that our usage policies prohibits and take action, for instance by filtering it.
+     *
+     * <p><strong>Request Body Schema</strong>
+     *
+     * <pre>{@code
+     * {
+     *     input (Required): [
+     *         String (Required)
+     *     ]
+     *     model: String (Optional)
+     * }
+     * }</pre>
+     *
+     * <p><strong>Response Body Schema</strong>
+     *
+     * <pre>{@code
+     * {
+     *     id (Required): String,
+     *     model (Required): String,
+     *     results: [
+     *       {
+     *         categories (Required): {
+     *           <category_type>: Boolean
+     *           ...
+     *         },
+     *         category_scores: {
+     *           <category_type>: Double
+     *           ...
+     *         },
+     *         flagged (Required): Boolean
+     *       }
+     *     ]
+     * }
+     * }</pre>
+     *
+     * Each category_type above is one of:
+     *
+     * hate: Content that expresses, incites, or promotes hate based on race, gender, ethnicity, religion,
+     * nationality, sexual orientation, disability status, or caste. Hateful content aimed at non-protected
+     * groups (e.g., chess players) is not covered by this category.
+     * hate/threatening: Hateful content that also includes violence or serious harm towards the targeted group.
+     * self-harm: Content that promotes, encourages, or depicts acts of self-harm, such as suicide, cutting,
+     * and eating disorders.
+     * sexual: Content meant to arouse sexual excitement, such as the description of sexual activity, or that
+     * promotes sexual services (excluding sex education and wellness).
+     * sexual/minors: Sexual content that includes an individual who is under 18 years old.
+     * violence: Content that promotes or glorifies violence or celebrates the suffering or humiliation of others.
+     * violence/graphic: Violent content that depicts death, violence, or serious physical injury in extreme
+     * graphic detail.
+     *
+     * @param modelId deployment id of the deployed model.
+     * @param moderationsOptions the configuration information for a moderation request.
+     * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws HttpResponseException thrown if the request is rejected by server.
+     * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
+     * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
+     * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the result of the moderation check to see if the input violates OpenAI's usage policies
+     * along with {@link Response} on successful completion of {@link Mono}.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<Response<BinaryData>> getModerationsWithResponseAsync(String modelId,
+        BinaryData moderationsOptions, RequestOptions requestOptions) {
+        final String accept = "application/json";
+
+        // OpenAI has model ID in request body
+        BinaryData moderationsOptionsUpdated = BinaryData.fromObject(
+            moderationsOptions.toObject(ModerationOptions.class)
+                .setModel(modelId)
+        );
+
+        return FluxUtil.withContext(
+            context ->
+                service.getModerations(
+                    OPEN_AI_ENDPOINT,
+                    accept,
+                    moderationsOptionsUpdated,
+                    requestOptions,
+                    context));
+    }
+
+
+    /**
+     * Gets moderations for the provided input. The moderations endpoint is a tool you can use to check
+     * whether content complies with OpenAI's usage policies. Developers can thus identify content
+     * that our usage policies prohibits and take action, for instance by filtering it.
+     *
+     * <p><strong>Request Body Schema</strong>
+     *
+     * <pre>{@code
+     * {
+     *     input (Required): [
+     *         String (Required)
+     *     ]
+     *     model: String (Optional)
+     * }
+     * }</pre>
+     *
+     * <p><strong>Response Body Schema</strong>
+     *
+     * <pre>{@code
+     * {
+     *     id (Required): String,
+     *     model (Required): String,
+     *     results: [
+     *       {
+     *         categories (Required): {
+     *           <category_type>: Boolean
+     *           ...
+     *         },
+     *         category_scores: {
+     *           <category_type>: Double
+     *           ...
+     *         },
+     *         flagged (Required): Boolean
+     *       }
+     *     ]
+     * }
+     * }</pre>
+     *
+     * Each category_type above is one of:
+     *
+     * hate: Content that expresses, incites, or promotes hate based on race, gender, ethnicity, religion,
+     * nationality, sexual orientation, disability status, or caste. Hateful content aimed at non-protected
+     * groups (e.g., chess players) is not covered by this category.
+     * hate/threatening: Hateful content that also includes violence or serious harm towards the targeted group.
+     * self-harm: Content that promotes, encourages, or depicts acts of self-harm, such as suicide, cutting,
+     * and eating disorders.
+     * sexual: Content meant to arouse sexual excitement, such as the description of sexual activity, or that
+     * promotes sexual services (excluding sex education and wellness).
+     * sexual/minors: Sexual content that includes an individual who is under 18 years old.
+     * violence: Content that promotes or glorifies violence or celebrates the suffering or humiliation of others.
+     * violence/graphic: Violent content that depicts death, violence, or serious physical injury in extreme
+     * graphic detail.
+     *
+     * @param modelId deployment id of the deployed model.
+     * @param moderationsOptions the configuration information for a moderation request.
+     * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws HttpResponseException thrown if the request is rejected by server.
+     * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
+     * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
+     * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the result of the moderation check to see if the input violates OpenAI's usage policies
+     * along with {@link Response} on successful completion of {@link Mono}.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Response<BinaryData> getModerationsWithResponse(String modelId, BinaryData moderationsOptions,
+                                                           RequestOptions requestOptions) {
+        final String accept = "application/json";
+
+        // OpenAI has model ID in request body
+        BinaryData moderationOptionsUpdated = BinaryData.fromObject(
+            moderationsOptions.toObject(ModerationOptions.class)
+                .setModel(modelId)
+        );
+
+        return service.getModerationsSync(
+            OPEN_AI_ENDPOINT,
+            accept,
+            moderationOptionsUpdated,
             requestOptions,
             Context.NONE);
     }